Refactor settings and queryapi pages
- Settings and Dashboard have been combined
- QueryAPI is no longer named Settings
