+++ Development/public/manifest.json
@@ -3,6 +3,6 @@
   "name": "nmos-js client",
   "start_url": "./index.html",
   "display": "standalone",
-  "theme_color": "#000000",
+  "theme_color": "#2196f3",
   "background_color": "#ffffff"
 }
